% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curlErrorHandler.R
\name{drive_upload_mod}
\alias{drive_upload_mod}
\title{Modified version of googledrive::drive_upload with error handling}
\usage{
drive_upload_mod(mediaInput, pathInput, NumberOfAttempts = 5)
}
\arguments{
\item{mediaInput}{media argument sent to \code{googledrive::drive_upload}}

\item{pathInput}{path argument sent to \code{googledrive::drive_upload}}

\item{NumberOfAttempts}{maximum number of attempts}
}
\value{
nothing is returned.
}
\description{
A curl error occasionally appears when using googledrive::drive_upload, but appears to be overcome by repeating the command.
}
\examples{
### error message:
# Error in curl::curl_fetch_memory(url, handle = handle) : 
#   Error in the HTTP2 framing layer


}
