% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_EDEN.R, R/subset_EDEN.R
\name{merge.eden}
\alias{merge.eden}
\alias{subset.eden}
\title{Merge EDEN data}
\usage{
\method{merge}{eden}(x, y = NULL, ...)

subset.eden(x, subset, ...)
}
\arguments{
\item{x}{EDEN object}

\item{y}{EDEN object}

\item{subset}{integer, character (in %Y-%m-%d format), or Date. Should indicate the layers (represented as integer or by the corresponding date)}
}
\value{
a single `eden` object, which is a list with two elements: (1) the date(s) used, and (2) a spatial object with water levels (centimeters relative to soil surface) in the EDEN grid.

a single `eden` object, which is a list with two elements: (1) the date(s) used, and (2) a spatial object with water levels (centimeters relative to soil surface) in the EDEN grid.
}
\description{
Merge EDEN objects (lists with data and dates)

Subset EDEN objects (lists with data and dates). Works only on SpatRaster data (raster data will be converted and returned as SpatRaster).
}
\examples{

\dontrun{
dat1 <- getEDEN(EDEN_date = Sys.Date(), quarterly = TRUE, returnType = 'terra')
dat2 <- getEDEN(EDEN_date = Sys.Date()-95, quarterly = TRUE, returnType = 'terra')

dat3 <- merge(x = dat1, y = dat2)

dat1$date
dat2$date
dat3$date

}


\dontrun{
dat1 <- getEDEN(EDEN_date = "2022-01-01", quarterly = TRUE, returnType = 'terra')

### subset by index
dat2 <- subset(x = dat1, subset = 5)

### subset by date
dat3 <- subset(x = dat1, subset = "2022-01-15")

### subset using date range
dateRange <- seq.Date(from = as.Date("2022-01-01"), 
                      to = as.Date("2022-01-15"), by = "day")
dat4 <- subset(x = dat1, subset = dateRange)

dat1$date
dat2$date
dat3$date
dat4$date

}

}
