% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_EDEN.R
\name{extract.eden}
\alias{extract.eden}
\title{Extract values from an EDEN object}
\usage{
extract.eden(x, y, fun = NULL, ...)
}
\arguments{
\item{x}{EDEN object}

\item{y}{SpatVector locations to extract data for}

\item{fun}{function to apply when extracted data have multiple cells. Untested with FUN = NULL.}
}
\value{
dataframe
}
\description{
Extract values from the SpatRaster portion of an EDEN object. This is a wrapper for terra::extract().
}
\examples{

\dontrun{
dat1 <- getEDEN(EDEN_date = "2022-01-01", quarterly = TRUE, returnType = 'terra')

sfwmd <- terra::vect(st_as_sf(EvergladesEBM::sfwmd))
wca1  <- sfwmd[sfwmd$NAME == "CONSERVATION AREA 1",]
wca2  <- sfwmd[sfwmd$NAME == "CONSERVATION AREA 2A",]
wca3  <- sfwmd[sfwmd$NAME == "CONSERVATION AREA 3A",]

extract.eden(x = dat1, y = wca1 + wca2 + wca3, fun = mean, na.rm = TRUE)

}

}
