% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire172.R
\docType{data}
\name{fire172}
\alias{fire172}
\title{Data: 2017 fire history shapefile}
\format{A shapefile of fires in 2017 in Big Cypress National Preserve and Everglades National Park}
\usage{
fire172
}
\description{
A shapefile of fires in 2017 occurring in Big Cypress National Preserve and Everglades National Park
}
\examples{

\dontrun{
summary(fire172)

### to create:
fireShps <- grep(x = list.files("C:/ESRIDATA/EVER_fireHistory", 
    full.names = TRUE), pattern = "shp$", value = TRUE)
fire17 <- st_read(fireShps[1])
fire18 <- st_read(fireShps[2])
fire19 <- st_read(fireShps[3])

st_transform(fire17, st_crs(a)$epsg)
st_transform(fire18, st_crs(a)$epsg)
st_transform(fire19, st_crs(a)$epsg)

fire17 <- fire17[, "OBJECTID"]
fire18 <- fire18[, "OBJECTID"] 
fire19 <- fire19[, "OBJECTID"]
fire17$burnYear <- "2017" 
fire18$burnYear <- "2018" 
fire19$burnYear <- "2019"

st_crs(fire17) <- st_crs(fire18) <- st_crs(fire19) <- st_crs(a)$epsg

fire172 <- sf::st_buffer(fire17, dist = 0)
fire182 <- sf::st_buffer(fire18, dist = 0)
fire192 <- sf::st_buffer(fire19, dist = 0)

save("fire172", "fire182", "fire192", file = "data/burnHistory.RData")

}

}
\keyword{burn}
\keyword{fire,}
\keyword{fuel}
\keyword{history,}
