% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_2020.R
\docType{data}
\name{fire_2020}
\alias{fire_2020}
\title{Data: 2020 fire history shapefile}
\format{
A shapefile of fires in 2020 in Big Cypress National Preserve and Everglades National Park
}
\usage{
fire_2020
}
\description{
A shapefile of fires in fiscal year 2020 (1 Oct 2019 - 30 Sept 2020) occurring in Big Cypress National Preserve and Everglades National Park
}
\examples{

\dontrun{
summary(fire_2020)

}

}
\keyword{burn}
\keyword{fire,}
\keyword{fuel}
\keyword{history,}
