% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_2022.R
\docType{data}
\name{fire_2022_cy}
\alias{fire_2022_cy}
\title{Data: 2022 fire history shapefile}
\format{
A shapefile of fires in 2022 in Big Cypress National Preserve and Everglades National Park
}
\usage{
fire_2022_cy
}
\description{
A shapefile of fires in calendar year 2022 (through 25 March 2022) occurring in Big Cypress National Preserve and Everglades National Park
}
\examples{

\dontrun{
fireShps <- grep(x = list.files("C:/ESRIDATA/EVER_fireHistory/20220328_fireHistory", 
recursive = TRUE, full.names = TRUE), 
pattern = "shp$", value = TRUE)
fire_2022_cy <- sf::st_read(grep(x = fireShps, pattern = "_2022", value = TRUE))

fire_2022_cy <- st_transform(fire_2022_cy, crs = sf::st_crs(fireHydro::edenDEM))
save(fire_2022_cy, file = "data/fire_2022.RData")
}

}
\keyword{burn}
\keyword{fire,}
\keyword{fuel}
\keyword{history,}
