% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnualEDEN.R
\name{getAnnualEDEN}
\alias{getAnnualEDEN}
\title{Downloads and imports archived EDEN data at annual or multi-annual time steps}
\usage{
getAnnualEDEN(
  years,
  DEM = raster(system.file("extdata/edenDEM.grd", package = "fireHydro"))
)
}
\arguments{
\item{years}{a vector specifying which years of EDEN data should be downloaded. Please consider RAM limitations - a year of data can be ~ 300 Mb of RAM. If this is an issue, consider using getAnnualEDEN() and do analysis/extraction incrementally in a loop.}

\item{DEM}{raster digital elevation model for south Florida. Used to subtract land elevations from water surface to get water depths. The default DEM is a USGS/EDEN product.  If `DEM = NULL`, output will be water surface in centimeters NAVD88.}
}
\value{
list \code{getAnnualEDEN} returns a list with two versions of the same data: (1) the dates used, and (2) a raster stack object with a layer for each day, containing water level data for the EDEN grid (units = cm rel. to soil surface).
}
\description{
Downloads and combines netCDF files from https://sofia.usgs.gov/eden/models/watersurfacemod_download.php, unzips and loads them into the workspace. Zip files are deleted after loading. This function makes `fireHydro` able to operate with complete independence from Department of Interior servers. This code generates a water depth map using the USGS water surface data and the USGS EDEN digital elevation map (present in this R package as raster layer "edenDEM").
}
\examples{

\dontrun{

edenDat <- getAnnualEDEN(years = 2019:2020)
}

}
