% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEDEN.R
\name{getEDEN}
\alias{getEDEN}
\title{Downloads and imports EDEN data}
\usage{
getEDEN(EDEN_date)
}
\arguments{
\item{EDEN_date}{EDEN date to be used for water levels. Should be an 8-digit numeric or character stirng, e.g., "20181018"}
}
\value{
sf \code{getEDEN} produces an sf shapefile.
}
\description{
Downloads GeoTiff files from https://sofia.usgs.gov/eden/models/real-time.php , unzips and loads them into the workspace. Zip files are deleted after loading. This function makes `fireHydro` able to operate with complete independence from Department of Interior servers.
}
\examples{

\dontrun{
a <- getEDEN(EDEN_date = "20181018")

### getEDEN output can then be used in getFireHydro
getFireHydro(EDEN_date = "20181018", 
     EDEN_GIS_directory = a,
     output_shapefile = NULL,
     fireSpreadExport = "fireRisk.pdf", waterLevelExport = "waterLevels.pdf")
}

}
