% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFireHydro.R
\name{getFireHydro}
\alias{getFireHydro}
\title{Generate fire-hydro shapefile}
\usage{
getFireHydro(EDEN_date, 
    output_shapefile = paste0(tempdir(), "/output_", EDEN_date, ".shp"), 
    waterLevelExport = NULL,
    fireSpreadExport = NULL,
    csvExport = NULL,
    EDEN_GIS_directory = "detect",
    vegetation_shp = fireHydro::vegetation,
    BICY_EVER_PlanningUnits_shp = fireHydro::BICY_EVER_PlanningUnits,
    returnShp = FALSE,
    figureWidth = 6,
    figureHeight = 4,
    ggBaseSize = 12,
    burnHist = FALSE)
}
\arguments{
\item{EDEN_date}{EDEN date to be used for water levels. Should be a character stirng, e.g., "20181018"}

\item{output_shapefile}{file address for shapefile output}

\item{waterLevelExport}{NULL or a character vector specifying the file address/name used for exporting an image file of water level categories (e.g., /home/waterLevels.pdf).}

\item{fireSpreadExport}{NULL or a character vector specifying the file address/name used for exporting an image file of fire spread risk (e.g., /home/fireSpreadRisk.pdf).}

\item{csvExport}{If an exported .csv file of the output is desired, include a file addess/name here (e.g., "fireHydroOutput.csv")}

\item{EDEN_GIS_directory}{The source for EDEN data. For users with access to the SFNRC's physical drive, the default value (\code{"detect"}) will identify the parent directory where EDEN water level data are located ("/opt/physical/gis/eden/" on linux; "Y:/gis/eden/" on Windows). This can alternative be the specific address of a shapefile of EDEN data. This can also be a character string naming an object in the working environment.}

\item{vegetation_shp}{shapefile of vegetation data in Big Cypress and Everglades}

\item{BICY_EVER_PlanningUnits_shp}{shapefile of polygons representing Big Cypress and Everglades planning units}

\item{returnShp}{TRUE/FALSE determinant of whether output is returned to the working environment}

\item{figureWidth}{width of output figure, in inches}

\item{figureHeight}{height of output figure, in inches}

\item{ggBaseSize}{base_size argument passed to ggplot theme.}

\item{burnHist}{logical, indicates whether burn history should be included to create a gradient of fire spread risk. If this is set to TRUE, the past three years of burn history will be used to identify zones of moderately low spread risk (high present-day risk but burned in current calendar year), moderate risk (high present-day risk but burned in previous calendar year) and moderately high risk (high present-day risk but burned two years ago).}
}
\value{
dataframe \code{getFireHydro} produces a shapefile.
}
\description{
Generates shapefile showing fire potential. Presently only works on the SFNRC network (for EDEN data access)
}
\examples{

\dontrun{
getFireHydro(EDEN_date = "20181018", 
     output_shapefile = NULL,
     fireSpreadExport = "fireRisk.png", waterLevelExport = "waterLevels.png")
}

}
