% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectionWarningHandler.R
\name{intersectionWarningHandler}
\alias{intersectionWarningHandler}
\title{Warning handler for sf::st_intersection}
\usage{
intersectionWarningHandler(warningMessage)
}
\arguments{
\item{warningMessage}{the warning message produced by \code{sf::st_intersection}}
}
\value{
nothing is returned.
}
\description{
Warning handler for sf::st_intersection
}
\examples{
#   dat <- data.frame(dat = c("5", "10", "-", "helloWorld"))
#   withCallingHandlers(
#     as.numeric(as.character(dat$dat)), warning = intersectionWarningHandler)
}
