% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_eden.R
\name{is.eden}
\alias{is.eden}
\title{Check for EDEN object}
\usage{
is.eden(x)
}
\arguments{
\item{x}{object to check}
}
\value{
logical
}
\description{
Checks object to see if it is an EDEN object
}
\examples{
edenList <- list(date = Sys.Date(), data = raster::raster())
is.eden(edenList)
}
