% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_EDEN.R
\name{merge.eden}
\alias{merge.eden}
\title{Merge EDEN lists}
\usage{
\method{merge}{eden}(x)
}
\arguments{
\item{x}{list of eden objects}
}
\value{
eden \code{getEDEN} returns an `eden` object, which is a list with two elements: (1) the date(s) used, and (2) a spatial object with water levels (centimeters relative to soil surface) in the EDEN grid.
}
\description{
Merge eden objects (lists with data and dates)
}
\examples{

\dontrun{
dat1 <- getEDEN(EDEN_date = Sys.Date(), quarterly = TRUE, returnType = 'raster')
dat2 <- getEDEN(EDEN_date = Sys.Date()-95, quarterly = TRUE, returnType = 'raster')

dat3 <- merge.eden(x = list(dat1, dat2))

dat1$date
dat2$date
dat3$date

}

}
