% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EDEN.R
\name{plot.eden}
\alias{plot.eden}
\title{Plot EDEN data}
\usage{
plot.eden(x, y = NULL, ...)
}
\arguments{
\item{x}{EDEN object}

\item{y}{NULL; not used}
}
\value{
a plot
}
\description{
Plot EDEN objects (lists with data and dates)
}
\examples{

\dontrun{
dat1 <- getEDEN(EDEN_date = Sys.Date(), returnType = 'terra')

plot(dat1)

dat2 <- getEDEN(EDEN_date = Sys.Date(), returnType = 'sf')
plot(dat2)
}

}
