% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_EDEN.R
\name{read.eden}
\alias{read.eden}
\title{Read EDEN data saved to disk}
\usage{
read.eden(filename)
}
\arguments{
\item{filename}{character. Input filename with or without a file extension; extension is ignored. two files are saved: (1) a geoTiff with the data, and (2) a text file with the dates. These are combined when loaded using `read.eden()`}
}
\value{
none
}
\description{
Load a saved EDEN object (list with data and dates). Returns an eden obeject with SpatRaster data.
}
\examples{

\dontrun{
dat1 <- getEDEN(EDEN_date = "2022-01-01", quarterly = TRUE, returnType = 'terra')

write.eden(x = dat1, filename = "test_dat1.tif")
newdat <- read.eden("test_dat1.tif")

newdat

}

}
