% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_EDEN.R
\name{write.eden}
\alias{write.eden}
\title{Write EDEN data to disk}
\usage{
write.eden(x, filename)
}
\arguments{
\item{x}{EDEN object}

\item{filename}{character. Output filename with or without a file extension; extension is ignored. two files are saved: (1) a geoTiff with the data, and (2) a text file with the dates. These are combined when loaded using `read.eden()`}
}
\value{
none
}
\description{
Write an EDEN object (list with data and dates) to a file. Works only on SpatRaster data (raster data will be converted to SpatRaster). Two files
}
\examples{

\dontrun{
dat1 <- getEDEN(EDEN_date = "2022-01-01", quarterly = TRUE, returnType = 'terra')

write.eden(x = dat1, filename = "test_dat1.tif")
newdat <- read.eden("test_dat1.tif")

newdat

}

}
