% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_EDEN.R
\name{merge.eden}
\alias{merge.eden}
\title{Merge EDEN data}
\usage{
\method{merge}{eden}(x, y, ...)
}
\arguments{
\item{x}{EDEN object}

\item{y}{EDEN object}
}
\value{
a single `eden` object, which is a list with two elements: (1) the date(s) used, and (2) a spatial object with water levels (centimeters relative to soil surface) in the EDEN grid.
}
\description{
Merge EDEN objects (lists with data and dates)
}
\examples{

\dontrun{
dat1 <- getEDEN(EDEN_date = Sys.Date(), quarterly = TRUE, returnType = 'terra')
dat2 <- getEDEN(EDEN_date = Sys.Date()-95, quarterly = TRUE, returnType = 'terra')

dat3 <- merge.eden(x = list(dat1, dat2))

dat1$date
dat2$date
dat3$date

}

}
