% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edenDEM.R
\docType{data}
\name{edenDEM}
\alias{edenDEM}
\title{Data: A raster DEM used to calculate EDEN water depths from water surfaces}
\format{
A raster layer. Downloaded from https://sofia.usgs.gov/eden/models/groundelevmod.php on 20190705. Metadata available here: http://sofia.usgs.gov/metadata/sflwww/eden_em_oc11.html
}
\usage{
edenDEM
}
\description{
A raster digital elevation map for the greater Everglades. Units = meters NAVD88.
}
\examples{
\dontrun{
### code used to generate object
edenDEM <- raster::raster(system.file("extdata/edenDEM.grd", package = "fireHydro"))
# writeRaster(edenDEM, file = paste0(here::here(), "/inst/extdata/edenDEM.grd"))
edenDEM <- edenDEM * 1 # store full raster in memory
save(edenDEM, file = "C:/RDATA/fireHydro/data/edenDEM.RData")
 
}

}
\keyword{DEM}
\keyword{digital}
\keyword{elevation}
\keyword{model}
