% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEDEN.R
\name{getEDEN}
\alias{getEDEN}
\title{Downloads and imports EDEN data}
\usage{
getEDEN(EDEN_date = gsub(Sys.Date(), pattern  = "-", replacement = ""), 
    exact = FALSE, 
    DEM = raster(system.file("extdata/edenDEM.grd", package = "fireHydro")))
}
\arguments{
\item{EDEN_date}{EDEN date to be used for water levels. Should be an 8-digit numeric or character stirng, e.g., "20181018". By default, today's date is used; if "exact = FALSE" this returns the most recent EDEN data available.}

\item{exact}{logical; if TRUE, output is only returned if the requested date is available. If exact = FALSE, the function responds to an invalid EDEN_date input by returning data from the most recent available date}

\item{DEM}{raster digital elevation model for south Florida. Used to subtract land elevations from water surface to get water depths. The default DEM is a USGS/EDEN product.}
}
\value{
list \code{getEDEN} returns a list with two elements: (1) the date used, and (2) an sf object with water levels in the EDEN grid.
}
\description{
Downloads GeoTiff files from https://sofia.usgs.gov/eden/models/real-time.php , unzips and loads them into the workspace. Zip files are deleted after loading. This function makes `fireHydro` able to operate with complete independence from Department of Interior servers. This code generates a water depth map using the USGS water surface data and the USGS EDEN digital elevation map (present in this R package as raster layer "edenDEM").
}
\examples{

\dontrun{
### default behavior is to return data from the most recent date
a <- getEDEN()

### getEDEN output can be directly used in getFireHydro()
a.fire <- getFireHydro(EDEN_date = a$date, 
     EDEN_GIS_directory = a$data,
     fireSpreadExport = paste0("fireRisk_", a$date, ".png"), 
     waterLevelExport = paste0("waterLevels_", a$date, ".png"))
     
### only the most recent data are available as individual geotiffs.
### Use 'exact = TRUE' to return NULL if an exact date match isn't found
edenDat.false <- getEDEN(EDEN_date = "20180101", exact = FALSE)
edenDat.true  <- getEDEN(EDEN_date = "20180101", exact = TRUE)

edenDat.false
edenDat.true 
# reporting the date with the output is particularly useful if the requested date is unavailable
}

}
