% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEDEN.R
\name{getEDEN}
\alias{getEDEN}
\title{Downloads and imports EDEN data}
\usage{
getEDEN(EDEN_date, exact = FALSE, 
    DEM = raster(system.file("extdata/edenDEM.grd", package = "fireHydro")))
}
\arguments{
\item{EDEN_date}{EDEN date to be used for water levels. Should be an 8-digit numeric or character stirng, e.g., "20181018"}

\item{exact}{logical; if TRUE, output is only returned if the requested date is available. If exact = FALSE, the function responds to an invalid EDEN_date input by returning data from the most recent available date}

\item{DEM}{raster digital elevation model for south Florida. Used to subtract land elevations from water surface to get water depths. The default DEM is a USGS/EDEN product.}
}
\value{
sf \code{getEDEN} returns an sf object.
}
\description{
Downloads GeoTiff files from https://sofia.usgs.gov/eden/models/real-time.php , unzips and loads them into the workspace. Zip files are deleted after loading. This function makes `fireHydro` able to operate with complete independence from Department of Interior servers. This code generates a water depth map using the USGS water surface data and the USGS EDEN digital elevation map (present in this R package as raster layer "edenDEM").
}
\examples{

\dontrun{
EDEN_date_target <- "20190606"
a <- getEDEN(EDEN_date = EDEN_date_target)

### getEDEN output can then be used in getFireHydro
a.fire <- getFireHydro(EDEN_date = EDEN_date_target, 
     EDEN_GIS_directory = "a",
     output_shapefile = NULL,
     fireSpreadExport = "fireRisk.png", waterLevelExport = "waterLevels.png", burnHist = TRUE)
     
### only the most recent data are available as individual geotiffs.
### Use 'exact = TRUE' to return NULL if an exact date match isn't found
edenDat.false <- getEDEN(EDEN_date = "20180101", exact = FALSE)
edenDat.true  <- getEDEN(EDEN_date = "20180101", exact = TRUE)

edenDat.false
edenDat.true
}

}
