% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFireHydro.R
\name{getFireHydro}
\alias{getFireHydro}
\title{Generate fire-hydro shapefile}
\usage{
getFireHydro(EDEN_date, 
output_shapefile = paste0(tempdir(), "/output_", EDEN_date, ".shp"), 
imageExport = NULL, csv = NULL,
EDEN_GIS_directory = "detect",
                         vegetation_shp = vegetation,
                         planningUnits_shp = planningUnits,
                         BICY_EVER_PlanningUnits_shp = BICY_EVER_PlanningUnits)
}
\arguments{
\item{EDEN_date}{EDEN date to be used for water levels. Should be a character stirng, e.g., "20181018"}

\item{output_shapefile}{file address for shapefile output}

\item{imageExport}{If an image output is desired, include a file addess/name here (e.g., "fireHydroOutput.png" or "fireHydroOutput.pdf").}

\item{csv}{If a .csv table of the output is desired, include a file addess/name here (e.g., "fireHydroOutput.csv")}

\item{EDEN_GIS_directory}{The source for EDEN data. For users with access to the SFNRC's physical drive, the default value (\code{"detect"}) will identify the parent directory where EDEN water level data are located ("/opt/physical/gis/eden/" on linux; "Y:/gis/eden/" on Windows). This can alternative be the specific address of a shapefile of EDEN data.}

\item{vegetation_shp}{shapefile of vegetation data in Big Cypress and Everglades}

\item{planningUnits_shp}{an sfc_POLYGON object representing Big Cypress and Everglades planning units. May be derived from BICY_EVER_PlanningUnits_shp}

\item{BICY_EVER_PlanningUnits_shp}{shapefile of polygons representing Big Cypress and Everglades planning units}
}
\value{
dataframe \code{getFireHydro} produces a shapefile.
}
\description{
Generates shapefile showing fire potential. Presently only works on the SFNRC network (for EDEN data access)
}
\examples{

\dontrun{
}

}
