% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFireHydro.R
\name{getFireHydro}
\alias{getFireHydro}
\title{Generate fire-hydro shapefile}
\usage{
getFireHydro(EDEN_date, 
    output_shapefile = paste0(tempdir(), "/output_", EDEN_date, ".shp"), 
    waterLevelExport = NULL,
    fireSpreadExport = NULL,
    csvExport = NULL,
    EDEN_GIS_directory = "detect",
    vegetation_shp = fireHydro::vegetation,
    BICY_EVER_PlanningUnits_shp = fireHydro::BICY_EVER_PlanningUnits,
    returnShp = FALSE,
    figureWidth = 6,
    figureHeight = 4,
    ggBaseSize = 12,
    burnHist = FALSE)
}
\arguments{
\item{EDEN_date}{EDEN date to be used for water levels. Should be a character stirng, e.g., "20181018"}

\item{output_shapefile}{file address for shapefile output}

\item{waterLevelExport}{NULL or a character vector specifying the file address/name used for exporting an image file of water level categories (e.g., /home/waterLevels.pdf).}

\item{fireSpreadExport}{NULL or a character vector specifying the file address/name used for exporting an image file of fire spread risk (e.g., /home/fireSpreadRisk.pdf).}

\item{csvExport}{If an exported .csv file of the output is desired, include a file addess/name here (e.g., "fireHydroOutput.csv")}

\item{EDEN_GIS_directory}{The source for EDEN data. Can be an \code{sf} object already in the working environment (such as the data output from \code{\link[fireHydro]{getEDEN}}) or, for users with access to the SFNRC's physical drive, the default value (\code{"detect"}) will identify the parent directory where EDEN water level data are located ("/opt/physical/gis/eden/" on linux; "Y:/gis/eden/" on Windows). This can alternative be the specific address of a shapefile of EDEN data. This can also be a character string naming an object in the working environment, such as that generated from getEDEN(). Setting this argument to refer to an sf object located in the working environment will result in an error.}

\item{vegetation_shp}{shapefile of vegetation data in Big Cypress and Everglades}

\item{BICY_EVER_PlanningUnits_shp}{shapefile of polygons representing Big Cypress and Everglades planning units}

\item{returnShp}{TRUE/FALSE determinant of whether output is returned to the working environment}

\item{figureWidth}{width of output figure, in inches}

\item{figureHeight}{height of output figure, in inches}

\item{ggBaseSize}{base_size argument passed to ggplot theme.}

\item{burnHist}{logical; if FALSE, fire spread risk is a binary variable (high/low); if TRUE, fire history during the preceding three years is used to split fire spread risk into a gradient of risk: High = high current fire spread risk and no burn history in past 3 years; Moderately High = high current fire spread risk and burned three years ago; Moderate = high current fire spread risk and burned two years ago; Moderately Low = high current fire spread risk and burned in the past year; Low = low current fire spread risk (regardless of burn history)}
}
\value{
sf
}
\description{
Generates shapefile showing fire potential. Presently only works on the SFNRC network (for EDEN data access)
}
\examples{

\dontrun{

getFireHydro(EDEN_date = "20181018", 
     output_shapefile = NULL,
     fireSpreadExport = "fireRisk.png", waterLevelExport = "waterLevels.png")

# save output in multiple file types
getFireHydro(EDEN_date = "20181018", 
     output_shapefile = NULL,
     fireSpreadExport = c("fireRisk.png", "fireRisk.pdf"))
     
# incorportate burn history to fire spread risk maps
getFireHydro(EDEN_date = "20181018", 
     output_shapefile = NULL,
     fireSpreadExport = c("fireRisk.png", "fireRisk.pdf"), burnHist = TRUE)

}

}
