% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOldEDEN.R
\name{getOldEDEN}
\alias{getOldEDEN}
\title{Downloads and imports archived EDEN data (from 2000 through the penultimate quarter in the record)}
\usage{
getOldEDEN(
  YYYYMMDD,
  returnType = "sf",
  baseURL = c("https://sofia.usgs.gov/eden/data/netcdf/v2/",
    "https://sofia.usgs.gov/eden/data/netcdf/v3/",
    "https://sofia.usgs.gov/eden/data/realtime2/"),
  urlEnding = c("v2prov", "v2prov_r2", "v2prov_r3", "v2r1", "v2r2", "v2r3", "v2rt_nc",
    "v3prov", "v3prov_r2", "v3prov_r3", "v3r1", "v3r2", "v3r3", "v3rt_nc", "v4prov",
    "v4prov_r2", "v4prov_r3", "v4r1", "v4r2", "v4r3", "v4rt_nc"),
  DEM = raster(system.file("extdata/edenDEM.grd", package = "fireHydro")),
  quarterly = FALSE
)
}
\arguments{
\item{YYYYMMDD}{EDEN date to be used for water levels. Should be an 8-digit numeric or character stirng, e.g., "20181018". By default, today's date is used; if "exact = FALSE" this returns the most recent EDEN data available.}

\item{returnType}{character; class of object returned. Acceptable options: "sf", "raster"}

\item{baseURL}{beginning part of url}

\item{urlEnding}{ending part of url}

\item{DEM}{raster digital elevation model for south Florida. Used to subtract land elevations from water surface to get water depths. The default DEM is a USGS/EDEN product. If `DEM = NULL`, output will be water surface in centimeters NAVD88.}

\item{quarterly}{logical; if set to TRUE, entire quarter is downloaded and returned as a RasterStack.}
}
\value{
eden \code{getOldEDEN} returns an `eden` object, which is a list with two elements: (1) the date used, and (2) a spatial object with water levels (centimeters relative to soil surface) in the EDEN grid.
}
\description{
Note: This function is deprecated. See \code{getQuarterlyEDEN} for a more generally useful function. Downloads netCDF files from https://sofia.usgs.gov/eden/models/watersurfacemod_download.php, unzips and loads them into the workspace. Zip files are deleted after loading. This function makes `fireHydro` able to operate with complete independence from Department of Interior servers. This code generates a water depth map using the USGS water surface data and the USGS EDEN digital elevation map (present in this R package as raster layer "edenDEM").
}
\examples{

\dontrun{

edenDat <- getOldEDEN(YYYYMMDD = "20170809", returnType = "raster")
}

}
