% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation.R
\docType{data}
\name{vegetation}
\alias{vegetation}
\title{Data: Vegetation shapefile for Big Cypress National Preserve and Everglades National Park}
\format{A shapefile of vegetation in Big Cypress National Preserve and Everglades National Park
\describe{
\item{NAME_SITE}{tbd}
\item{NAME_STATE}{tbd}
\item{Veg_Cat}{tbd}
\item{Notes}{tbd}
\item{Shape_Leng}{tbd}
\item{Shape_Area}{tbd}
\item{geometry}{tbd}
\item{FuelType}{tbd}
}}
\usage{
vegetation
}
\description{
Data: Vegetation shapefile for Big Cypress National Preserve and Everglades National Park
}
\keyword{fuel}
\keyword{vegetation,}
